requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.43.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
gunicorn==20.0.4
